Split DSODatabase/DSODatabaseBuilder files
