+++ src/celengine/dsodb.h
@@ -13,31 +13,28 @@
 #pragma once
 
 #include <cstdint>
-#include <iosfwd>
 #include <memory>
 #include <string>
 #include <string_view>
 #include <vector>
 
-#include <Eigen/Core>
-#include <Eigen/Geometry>
+#include "astroobj.h"
+#include "dsooctree.h"
 
-#include <celcompat/filesystem.h>
-#include <celengine/dsooctree.h>
-#include <celengine/name.h>
-
-constexpr inline unsigned int MAX_DSO_NAMES = 10;
+class DeepSkyObject;
+class DSODatabaseBuilder;
+class NameDatabase;
 
 // 100 Gly - on the order of the current size of the universe
 constexpr inline float DSO_OCTREE_ROOT_SIZE = 1.0e11f;
-
-class DSODatabaseBuilder;
 
 //NOTE: this one and starDatabase should be derived from a common base class since they share lots of code and functionality.
 class DSODatabase
 {
 public:
-    ~DSODatabase() = default;
+    static constexpr inline unsigned int MAX_DSO_NAMES = 10;
+
+    ~DSODatabase();
 
     DeepSkyObject* getDSO(const std::uint32_t) const;
     std::uint32_t size() const;
@@ -48,18 +45,6 @@ public:
     void getCompletion(std::vector<std::string>&, std::string_view) const;
 
     const celestia::engine::DSOOctree& getOctree() const;
-/*
-    void findVisibleDSOs(DSOHandler& dsoHandler,
-                         const Eigen::Vector3d& obsPosition,
-                         const Eigen::Quaternionf& obsOrientation,
-                         float fovY,
-                         float aspectRatio,
-                         float limitingMag) const;
-
-    void findCloseDSOs(DSOHandler& dsoHandler,
-                       const Eigen::Vector3d& obsPosition,
-                       float radius) const;
-*/
 
     std::string getDSOName    (const DeepSkyObject*, bool i18n = false) const;
     std::string getDSONameList(const DeepSkyObject*, const unsigned int maxNames = MAX_DSO_NAMES) const;
@@ -98,16 +83,3 @@ DSODatabase::getOctree() const
 {
     return m_octree;
 }
-
-class DSODatabaseBuilder
-{
-public:
-    bool load(std::istream&, const fs::path& resourcePath = fs::path());
-    std::unique_ptr<DSODatabase> build();
-
-private:
-    float calcAvgAbsMag();
-
-    std::vector<std::unique_ptr<DeepSkyObject>> m_dsos;
-    std::vector<std::pair<AstroCatalog::IndexNumber, std::string>> m_names;
-};
